scripts: do not update resolv.conf file if it is symlink #PSBM-146148
If resolv.conf is symlink so it updates under other service and
vzctl does not need to apply changes in it directly.

+++ dists/scripts/set_dns.sh
@@ -64,7 +64,7 @@ set_resolvconf()
 			put_param2 "${cfgfile}" search "${search}"
 		fi
 	fi
-	if [ -n "${server}" ]; then
+	if [ -n "${server}" -a ! -L "${cfgfile}" ]; then
 		[ -f "${cfgfile}" ] || touch "${cfgfile}"
 		${CP} ${cfgfile} ${cfgfile}.$$ || error "Can't copy file $cfgfile" $VZ_FS_NO_DISK_SPACE
 		sed "/nameserver.*/d" ${cfgfile} > ${cfgfile}.$$
